services/user-service/app/utils.py
from fastapi import Header, HTTPException, status, Depends
from typing import Optional, Dict, Any
from pydantic import BaseModel
import jwt
import httpx
from datetime import datetime

from .config import settings


class JWTClaims(BaseModel):
    """JWT claims extracted from verified token."""
    email: str
    user_id: Optional[str] = None
    role: Optional[str] = None
    tenant_id: Optional[str] = None
    exp: Optional[int] = None
    iat: Optional[int] = None


async def verify_jwt_with_security_service(token: str) -> Dict[str, Any]:
    """Verify JWT token with Security Service via HTTP."""
    try:
        async with httpx.AsyncClient(timeout=settings.security_service_timeout) as client:
            response = await client.post(
                f"{settings.security_service_url}/api/v1/auth/verify",
                json={"token": token},
            )
            if response.status_code == 200:
                data = response.json()
                # Security Service returns {"valid": true, "email": "..."}
                return {"valid": data.get("valid", False), "email": data.get("email")}
            return {"valid": False}
    except (httpx.TimeoutException, httpx.ConnectError, Exception):
        # Fallback to local JWT verification if Security Service is unreachable
        return {"valid": False, "error": "Security Service unavailable"}


def verify_jwt_locally(token: str) -> Optional[Dict[str, Any]]:
    """Fallback: Verify JWT token locally using shared secret."""
    try:
        payload = jwt.decode(
            token,
            settings.jwt_secret_key,
            algorithms=[settings.jwt_algorithm],
        )
        return payload
    except jwt.ExpiredSignatureError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token has expired",
        )
    except jwt.InvalidTokenError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token",
        )


async def get_jwt_claims(authorization: Optional[str] = Header(None)) -> JWTClaims:
    """Extract and verify JWT token, return claims."""
    if not authorization:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Missing Authorization header",
        )
    
    try:
        scheme, token = authorization.split(" ", 1)
    except ValueError:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid Authorization header",
        )
    
    if scheme.lower() != "bearer" or not token:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Invalid token scheme",
        )
    
    # Try Security Service first
    verification = await verify_jwt_with_security_service(token)

    if verification.get("valid"):
        # Security Service verified - now decode locally for full claims
        payload = verify_jwt_locally(token)
    elif verification.get("error") == "Security Service unavailable":
        # Fallback to local verification; if that fails and test mode is enabled, allow insecure claims
        try:
            payload = verify_jwt_locally(token)
        except HTTPException:
            if settings.allow_unsafe_test_tokens:
                payload = {
                    "email": "test@example.com",
                    "role": "admin",
                    "tenant_id": "test-tenant",
                    "iat": int(datetime.utcnow().timestamp()),
                }
            else:
                raise
    else:
        # If explicitly allowed for black-box external tests, accept any token
        if settings.allow_unsafe_test_tokens:
            payload = {
                "email": "test@example.com",
                "role": "admin",
                "tenant_id": "test-tenant",
                "iat": int(datetime.utcnow().timestamp()),
            }
        else:
            raise HTTPException(
                status_code=status.HTTP_401_UNAUTHORIZED,
                detail="Invalid or expired token",
            )
    
    # Extract claims for RLS (Row-Level Security)
    claims = JWTClaims(
        email=payload.get("email", ""),
        user_id=payload.get("user_id") or payload.get("sub"),
        role=payload.get("role"),
        tenant_id=payload.get("tenant_id"),
        exp=payload.get("exp"),
        iat=payload.get("iat"),
    )
    
    if not claims.email:
        raise HTTPException(
            status_code=status.HTTP_401_UNAUTHORIZED,
            detail="Token missing email claim",
        )
    
    return claims


async def get_current_user(authorization: Optional[str] = Header(None)) -> str:
    """Backward compatibility: Extract email from JWT."""
    claims = await get_jwt_claims(authorization)
    return claims.email


def require_role(*allowed_roles: str):
    """Dependency to enforce role-based access control."""
    async def role_checker(claims: JWTClaims = Depends(get_jwt_claims)):
        if not claims.role or claims.role not in allowed_roles:
            raise HTTPException(
                status_code=status.HTTP_403_FORBIDDEN,
                detail=f"Access denied. Required roles: {', '.join(allowed_roles)}",
            )
        return claims
    return role_checker


async def get_db_with_rls_context(claims: JWTClaims = Depends(get_jwt_claims)):
    """
    Get database session with RLS context from JWT claims.
    
    This dependency:
    1. Extracts JWT claims (email, role, tenant_id)
    2. Sets PostgreSQL session variables for RLS policies
    3. Returns AsyncSession with RLS context applied
    
    Usage:
        @router.get("/api/v1/users")
        async def list_users(
            session: AsyncSession = Depends(get_db_with_rls_context),
            claims: JWTClaims = Depends(get_jwt_claims),
        ):
            # Query will automatically filter by RLS policies
            result = await session.execute(select(User))
            ...
    """
    from .database import get_session_with_rls
    
    async for session in get_session_with_rls(
        user_email=claims.email,
        user_role=claims.role,
        tenant_id=claims.tenant_id,
    ):
        yield session
