requirements.txt
bcrypt>=4.1.2
PyJWT>=2.8.0
cryptography>=41.0.0
slowapi==0.1.9
# Core testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0

# Code quality and linting
black>=23.0.0
isort>=5.12.0
pylint>=2.17.0
flake8>=6.0.0
mypy>=1.0.0

# Testing utilities
requests>=2.28.0
httpx>=0.24.0
fastapi>=0.100.0
uvicorn>=0.23.0

# Development tools
pre-commit>=3.0.0
gitpython>=3.1.0

# Documentation
mkdocs>=1.4.0
mkdocs-material>=9.0.0

# Security scanning
bandit>=1.7.0
safety>=2.3.0